+++ CHANGELOG.md
@@ -17,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update TextMeshPro to 2.0.0
 
 ### Fixes
-
+- Fix slow sliders that were too slow because were processing to often.
+- Fix Color on Edit Diffuse. Old bug.
 
 ### To Implement
 - Windows Copy / Paste